pyproject.toml
[project]
name = "mcp-finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytest>=8.3.5",
    "requests>=2.32.3",
]
